PhosphoExperiment object accessors

These are methods for getting for setting accessors of PhosphoExperiment object. This provides some convenience for users.

UniprotID(x, ...)

UniprotID(x) <- value

GeneSymbol(x, ...)

GeneSymbol(x) <- value

Site(x, ...)

Site(x) <- value

Residue(x, ...)

Residue(x) <- value

Sequence(x, ...)

Sequence(x) <- value

Localisation(x, ...)

Localisation(x) <- value

# S4 method for PhosphoExperiment
UniprotID(x, withDimnames = TRUE)

# S4 method for PhosphoExperiment
GeneSymbol(x, withDimnames = TRUE)

# S4 method for PhosphoExperiment
Site(x, withDimnames = TRUE)

# S4 method for PhosphoExperiment
Residue(x, withDimnames = TRUE)

# S4 method for PhosphoExperiment
Sequence(x, withDimnames = TRUE)

# S4 method for PhosphoExperiment
Localisation(x, withDimnames = TRUE)

# S4 method for PhosphoExperiment
UniprotID(x) <- value

# S4 method for PhosphoExperiment
GeneSymbol(x) <- value

# S4 method for PhosphoExperiment
Site(x) <- value

# S4 method for PhosphoExperiment
Residue(x) <- value

# S4 method for PhosphoExperiment
Sequence(x) <- value

# S4 method for PhosphoExperiment
Localisation(x) <- value

Arguments

x

A PhosphoExperiment object to be assigned to.

...

Ignored for accessors.

value

A vector of values to set to respective accessor. See section Available methods for more details.

withDimnames

A logical(1), indicating whether the names of the vector should be applied.

Available methods

In the following code snippets, ppe is a PhosphoExperiment object.

UniprotID(ppe), UniprotID(ppe) <- value:

Get or set a Uniprot ID, where value is a character vector

GeneSymbol(ppe), GeneSymbol(ppe) <- value:

Get or set a gene symbol , where value is a character vector

Site(ppe), Site(ppe) <- value:

Get or set a phosphorylation site, where value is a numeric vector

Residue(ppe), Residue(ppe) <- value:

Get or set a residue of phosphorylation site, where value is a character

Sequence(ppe), Sequence(ppe) <- value:

Get or set a sequence, where value is a character vector

Localisation(ppe), Localisation(ppe) <- localisation:

Get or set a localisation score, where localisation is a numeric vector

Author

Taiyun Kim

Examples

example(PhosphoExperiment, echo = FALSE)

UniprotID(phosData) <- uniprot
head(UniprotID(phosData))
#>     Q6AYR1;TFG;S198;MSAFGLTDDQVSGPPSAPTEDRSGTPDSIAS 
#>                                               "TFG" 
#>   D3ZRN2;MED1;T1035;STGGSKSPGSSGRCQTPPGVATPPIPKITIQ 
#>                                              "MED1" 
#>    D3ZUD5;OFD1;S780;SSSPCLDRPSESPAASPTPCPERTPQPSSVP 
#>                                              "OFD1" 
#>  Q68FR3;INTS12;S127;DVPKKPRLEKPETRSSPITVQTSKDLAMADL 
#>                                            "INTS12" 
#>  B5DF98;MAP3K3;S175;PRSRHLSVSSQNPGRSSPPPGYVPERQQHIA 
#>                                            "MAP3K3" 
#> D3ZPU4;ERCC6L2;S913;RVPKNPICCKLLLGESEDSETEDPVKVNHDD 
#>                                           "ERCC6L2" 

GeneSymbol(phosData) <- symbol
head(GeneSymbol(phosData))
#>     Q6AYR1;TFG;S198;MSAFGLTDDQVSGPPSAPTEDRSGTPDSIAS 
#>                                               "TFG" 
#>   D3ZRN2;MED1;T1035;STGGSKSPGSSGRCQTPPGVATPPIPKITIQ 
#>                                              "MED1" 
#>    D3ZUD5;OFD1;S780;SSSPCLDRPSESPAASPTPCPERTPQPSSVP 
#>                                              "OFD1" 
#>  Q68FR3;INTS12;S127;DVPKKPRLEKPETRSSPITVQTSKDLAMADL 
#>                                            "INTS12" 
#>  B5DF98;MAP3K3;S175;PRSRHLSVSSQNPGRSSPPPGYVPERQQHIA 
#>                                            "MAP3K3" 
#> D3ZPU4;ERCC6L2;S913;RVPKNPICCKLLLGESEDSETEDPVKVNHDD 
#>                                           "ERCC6L2" 

Site(phosData) <- site
head(Site(phosData))
#>     Q6AYR1;TFG;S198;MSAFGLTDDQVSGPPSAPTEDRSGTPDSIAS 
#>                                                 198 
#>   D3ZRN2;MED1;T1035;STGGSKSPGSSGRCQTPPGVATPPIPKITIQ 
#>                                                1035 
#>    D3ZUD5;OFD1;S780;SSSPCLDRPSESPAASPTPCPERTPQPSSVP 
#>                                                 780 
#>  Q68FR3;INTS12;S127;DVPKKPRLEKPETRSSPITVQTSKDLAMADL 
#>                                                 127 
#>  B5DF98;MAP3K3;S175;PRSRHLSVSSQNPGRSSPPPGYVPERQQHIA 
#>                                                 175 
#> D3ZPU4;ERCC6L2;S913;RVPKNPICCKLLLGESEDSETEDPVKVNHDD 
#>                                                 913 

Residue(phosData) <- res
head(Residue(phosData))
#>     Q6AYR1;TFG;S198;MSAFGLTDDQVSGPPSAPTEDRSGTPDSIAS 
#>                                                 "S" 
#>   D3ZRN2;MED1;T1035;STGGSKSPGSSGRCQTPPGVATPPIPKITIQ 
#>                                                 "T" 
#>    D3ZUD5;OFD1;S780;SSSPCLDRPSESPAASPTPCPERTPQPSSVP 
#>                                                 "S" 
#>  Q68FR3;INTS12;S127;DVPKKPRLEKPETRSSPITVQTSKDLAMADL 
#>                                                 "S" 
#>  B5DF98;MAP3K3;S175;PRSRHLSVSSQNPGRSSPPPGYVPERQQHIA 
#>                                                 "S" 
#> D3ZPU4;ERCC6L2;S913;RVPKNPICCKLLLGESEDSETEDPVKVNHDD 
#>                                                 "S" 

Sequence(phosData) <- seq
head(Sequence(phosData))
#>     Q6AYR1;TFG;S198;MSAFGLTDDQVSGPPSAPTEDRSGTPDSIAS 
#>                   "MSAFGLTDDQVSGPPSAPTEDRSGTPDSIAS" 
#>   D3ZRN2;MED1;T1035;STGGSKSPGSSGRCQTPPGVATPPIPKITIQ 
#>                   "STGGSKSPGSSGRCQTPPGVATPPIPKITIQ" 
#>    D3ZUD5;OFD1;S780;SSSPCLDRPSESPAASPTPCPERTPQPSSVP 
#>                   "SSSPCLDRPSESPAASPTPCPERTPQPSSVP" 
#>  Q68FR3;INTS12;S127;DVPKKPRLEKPETRSSPITVQTSKDLAMADL 
#>                   "DVPKKPRLEKPETRSSPITVQTSKDLAMADL" 
#>  B5DF98;MAP3K3;S175;PRSRHLSVSSQNPGRSSPPPGYVPERQQHIA 
#>                   "PRSRHLSVSSQNPGRSSPPPGYVPERQQHIA" 
#> D3ZPU4;ERCC6L2;S913;RVPKNPICCKLLLGESEDSETEDPVKVNHDD 
#>                   "RVPKNPICCKLLLGESEDSETEDPVKVNHDD" 

Localisation(phosData) <- rnorm(nrow(phosData))
head(Localisation(phosData))
#>     Q6AYR1;TFG;S198;MSAFGLTDDQVSGPPSAPTEDRSGTPDSIAS 
#>                                          1.05455439 
#>   D3ZRN2;MED1;T1035;STGGSKSPGSSGRCQTPPGVATPPIPKITIQ 
#>                                         -1.03665796 
#>    D3ZUD5;OFD1;S780;SSSPCLDRPSESPAASPTPCPERTPQPSSVP 
#>                                         -0.03837552 
#>  Q68FR3;INTS12;S127;DVPKKPRLEKPETRSSPITVQTSKDLAMADL 
#>                                         -0.22830366 
#>  B5DF98;MAP3K3;S175;PRSRHLSVSSQNPGRSSPPPGYVPERQQHIA 
#>                                          0.18994387 
#> D3ZPU4;ERCC6L2;S913;RVPKNPICCKLLLGESEDSETEDPVKVNHDD 
#>                                         -0.01242549