Kinase substrate scoring

This function generates substrate scores for kinases that pass filtering based on both motifs and dynamic profiles

kinaseSubstrateScore(
    substrate.list,
    mat,
    seqs,
    numMotif = 5, 
    numSub = 1, 
    verbose = TRUE
)

Arguments

substrate.list
a list of kinases with each element containing an array of substrates.

mat
a matrix with rows correspond to phosphosites and columns correspond to samples.

seqs
an array containing aa sequences surrounding each of all phosphosites. Each sequence has length of 15 (-7, p, +7).

numMotif
minimum number of sequences used for compiling motif for each kinase. Default is 5.

numSub
minimum number of phosphosites used for compiling phosphorylation profile for each kinase. Default is 1.

verbose
Default to TRUE to show messages during the progress. All messages will be suppressed if set to FALSE

Value

A list of 4 elements. motifScoreMatrix, profileScoreMatrix, combinedScoreMatrix, ksActivityMatrix (kinase activity matrix) and their weights.

Examples

data('phospho.L6.ratio.pe')

#> Warning: data set ‘phospho.L6.ratio.pe’ not found

data('SPSs')
data('PhosphoSitePlus')

grps = gsub('_.+', '', colnames(phospho.L6.ratio.pe))

# Construct a design matrix by condition
design = model.matrix(~ grps - 1)

# phosphoproteomics data normalisation using RUV
L6.sites = paste(sapply(phospho.L6.ratio.pe@GeneSymbol, function(x)paste(x)),
                 ";",
                 sapply(phospho.L6.ratio.pe@Residue, function(x)paste(x)),
                 sapply(phospho.L6.ratio.pe@Site, function(x)paste(x)),
                 ";", sep = "")
ctl = which(L6.sites %in% SPSs)
phospho.L6.ratio.RUV = RUVphospho(phospho.L6.ratio.pe@assays@data$Quantification, 
                                  M = design, k = 3,ctl = ctl)

phosphoL6 = phospho.L6.ratio.RUV
rownames(phosphoL6) <- rownames(phospho.L6.ratio.pe)

# filter for up-regulated phosphosites
phosphoL6.mean <- meanAbundance(phosphoL6, grps = grps)
aov <- matANOVA(mat=phosphoL6, grps = grps)
phosphoL6.reg <- phosphoL6[(aov < 0.05) &
                        (rowSums(phosphoL6.mean > 0.5) > 0), drop = FALSE]
L6.phos.std <- standardise(phosphoL6.reg)

#> Error in apply(mat, 1, mean): dim(X) must have a positive length

idx <- match(rownames(L6.phos.std), L6.sites)

#> Error in rownames(L6.phos.std): object 'L6.phos.std' not found

L6.phos.seq <- phospho.L6.ratio.pe@Sequence[idx]

#> Error in eval(expr, envir, enclos): object 'idx' not found

L6.matrices <- kinaseSubstrateScore(PhosphoSite.mouse, L6.phos.std,
    L6.phos.seq, numMotif = 5, numSub = 1)

#> Error in rownames(mat): object 'L6.phos.std' not found